pyproject.toml
[tool.poetry]
name = "poetry-overfero"
version = "0.1.0"
description = ""
authors = ["Fehru Mandala Putra <fehruputramen22@gmail.com>"]
readme = "README.md"
packages = [{include = "poetry-overfero"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "~=2.2.0"
hydra-core = "~=1.3.0"
pydantic = "~=2.6.0"
transformers = {version = "~=4.39.0", extras = ["tensorflow"]}
fsspec = {version = "~=2024.3.0", extras = ["gcs"]}
gcsfs = "~=2024.3.0"
tensorflow = "~=2.16.0"
mlflow = "~=2.11.0"
psycopg2-binary = "~=2.9.0"
google-cloud-compute = "~=1.18.0"
google-cloud-secret-manager = "~=2.18.0"


[tool.poetry.group.dev.dependencies]
jupyterlab = "~=4.1.0"
pytest = "~=8.0.0"
black = "~=24.2.0"
isort = "~=5.13.0"
flake8 = "~=7.0.0"
mypy = "~=1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party = ["poetry-overfero"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.10"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true

disallow_subclassing_any = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0